adding add to queue support

+++ src/fossegrim.lib/Models/PlaylistItem.cs
@@ -13,8 +13,10 @@ public class PlaylistItem
     public Guid MediaItemId { get; set; }
     public MediaItem? MediaItem { get; set; }
 
-    // Position of this track within the playlist, used for ordering
-    public int Position { get; set; }
+    // Fractional position of this track within the playlist, used for ordering.
+    // Moving an item only needs to set a value between its new neighbors,
+    // so a drag-and-drop reorder writes a single row instead of the whole list.
+    public double Position { get; set; }
 
     public DateTime? DateAdded { get; set; }
 }
